video4 - stopped at 32:00, cartridge programCounter overflow fixed

+++ olc6502.h
@@ -30,7 +30,7 @@ public:
 	uint8_t      x = 0x00;    // X Register
 	uint8_t      y = 0x00;    // Y Register
 	uint8_t   stkp = 0x00;    // Stack Pointer (points to location in bus)
-	uint8_t     pc = 0x0000;  // Program Counter
+	uint16_t    pc = 0x0000;  // Program Counter
 	uint8_t status = 0x00;    // Status Register
 
 	void ConnectBus(Bus *n) { bus = n; }
